+++ DVSTracker/particle.h
@@ -6,13 +6,13 @@
 class Particle
 {
 public:
-    Particle(float posX = 0, float posY = 0, float uncert = FLT_MAX, int w = 0, double ts = 0);
-    void set(float posX, float posY, float uncert, int w, double ts);
+    Particle(float posX = 0, float posY = 0, float uncert = FLT_MAX, float w = 0, double ts = 0);
+    void set(float posX, float posY, float uncert, float w, double ts);
     void reset();
 
     float x,y;
     float uncertainty;
-    int weight;
+    float weight;
     double timeStamp;
 };
